// src/features/filter/types/filter-type.ts
export enum OrderFilterKey {
  status = 'status',
  docNumber = 'docNumber',
  amount = 'amount',
  docType = 'docType',
  recipientInn = 'recipientInn',
  recipientKpp = 'recipientKpp',
  recipientPersonalAccount = 'recipientPersonalAccount',
  recipientName = 'recipientName',
  recipientBankAccount = 'recipientBankAccount',
  recipientBankName = 'recipientBankName',
  payerInn = 'payerInn',
  payerKpp = 'payerKpp',
  payerPersonalAccount = 'payerPersonalAccount',
  payerName = 'payerName',
  payerBankAccount = 'payerBankAccount',
  payerBankName = 'payerBankName',
  docDateFrom = 'docDateFrom',
  docDateTo = 'docDateTo',
  createdTsFrom = 'createdTsFrom',
  createdTsTo = 'createdTsTo',
  createdTs = 'createdTs',
  primaryDocCode = 'primaryDocCode',
  typeOfOperation = 'typeOfOperation',
  code = 'code',
}

export type FilterType = {
  [OrderFilterKey.docNumber]: null | string,
  [OrderFilterKey.docDateFrom]: null | string,
  [OrderFilterKey.docDateTo]: null | string,
  [OrderFilterKey.createdTsFrom]: null | string,
  [OrderFilterKey.createdTsTo]: null | string,
  [OrderFilterKey.amount]: null | number,
  [OrderFilterKey.docType]: null | string,
  [OrderFilterKey.primaryDocCode]: null | string,
  [OrderFilterKey.code]: null | string,
  [OrderFilterKey.typeOfOperation]: null | string,
  [OrderFilterKey.recipientInn]: null | string,
  [OrderFilterKey.recipientKpp]: null | string,
  [OrderFilterKey.recipientPersonalAccount]: null | string,
  [OrderFilterKey.recipientName]: null | string,
  [OrderFilterKey.recipientBankAccount]: null | string,
  [OrderFilterKey.recipientBankName]: null | string,
  [OrderFilterKey.payerInn]: null | string,
  [OrderFilterKey.payerKpp]: null | string,
  [OrderFilterKey.payerPersonalAccount]: null | string,
  [OrderFilterKey.payerName]: null | string,
  [OrderFilterKey.payerBankAccount]: null | string,
  [OrderFilterKey.payerBankName]: null | string,
  [OrderFilterKey.status]: {
    label: string,
    value: string,
  }[] | null,
}